replace word
In general I struggle with the word of democratisation of technology. I
think de-democratisation is even worse. The direct opposite is
autocratisation (when we talk about change of political systems). Let's
call it what it really means, democratisation of technologt means making
technology more accessible (especially to non-specialists). In this case
it is making research less accessible.